Changed match rules to allow reconnections
Previously, we matched /org/bluez/hci0/<MAC>/fd0 and fd1, but we
actually want to monitor /org/bluez/hci0/<MAC>/fd* because the number
after fd is incremented on reconnect. Therefore we switch from mathing
path to matching path_namespace. This was originally avoided as it has
the downside of also matching /org/bluez/hci0/<MAC>/sep* and /player*,
but I'm not aware of any matching signals that these endpoints would
emit. Therefore we're gonna give this a try